feat: little refactor
