+++ apps/api/package.json
@@ -21,7 +21,8 @@
     "pg-hstore": "^2.3.4",
     "rand-token": "^1.0.1",
     "sequelize": "^6.35.2",
-    "cache-manager": "^5.4.0"
+    "cache-manager": "^5.4.0",
+    "sequelize-simple-cache": "^1.3.4"
   },
   "devDependencies": {
     "@diary-spo/shared": "workspace:*",
