+++ apps/web/.env.development
@@ -1,7 +1,9 @@
-VITE_SERVER_URLS=http://localhost:3003,http://127.0.0.1:3003,https://api.spo-diary.ru/v2,https://diary.kopchan7.keenetic.link/v2,https://diary.zm.noho.st/v2
-
+VITE_SERVER_URLS=https://api.spo-diary.ru/v2,https://diary.kopchan7.keenetic.link/v2,https://diary.zm.noho.st/v2
+# http://localhost:3003,http://127.0.0.1:3003,
 VITE_MODE=dev
 VITE_NODE_ENV=development
+
+VITE_ADMIN_PAGE_URL=http://localhost:5174/
 
 VITE_BETA_VERSION=2.0.2
 # main 51740302
