+++ apps/web/src/config/env.ts
@@ -1,6 +1,6 @@
 export const BASE_URLS = import.meta.env.VITE_SERVER_URLS?.split(',')
 export const MODE = import.meta.env.VITE_MODE
-
+export const ADMIN_PAGE = import.meta.env.VITE_ADMIN_PAGE_URL
 export const BETA_VERSION = import.meta.env.VITE_BETA_VERSION
 
 export const Mode = {
