+++ apps/web/src/methods/makeRequest.ts
@@ -1,6 +1,6 @@
 import { BASE_URLS } from '@config'
 import { HTTP_STATUSES, ServerResponse } from '@types'
-import axios, { AxiosResponse } from 'axios'
+import axios, { AxiosError, AxiosResponse } from 'axios'
 
 const makeRequest = async <T>(
   route: string,
@@ -16,24 +16,28 @@ const makeRequest = async <T>(
 
   for (const BASE_URL of BASE_URLS) {
     try {
-      const response = await axios({
+      const response = await axios<T>({
         method,
         url: `${BASE_URL}${route}`,
         headers,
         data: body,
-        timeout: 3000
+        timeout: 2000
       })
-
-      /** В случае ошибки авторизации мы не делаем запрос на другой сервер, а сразу возвращаем ответ **/
-      if (response.status === HTTP_STATUSES.UNAUTHORIZED) {
-        return response
-      }
 
       console.info('%c [makeRequest]', 'color: blueviolet', response)
 
-      return (await response.data()) as T
+      return response.data
     } catch (err) {
       console.info('%c [makeRequest]', 'color: blueviolet', err)
+
+      if (!(err instanceof AxiosError)) {
+        return
+      }
+
+      /** В случае ошибки авторизации мы не делаем запрос на другой сервер, а сразу возвращаем ответ **/
+      if (err?.response?.status === HTTP_STATUSES.UNAUTHORIZED) {
+        return err.response
+      }
     }
   }
